Bug Fix 11
- Fix for paused screen breaking users
