+++ scripts/helpers/generic-utils.mjs
@@ -120,3 +120,18 @@ export function getConfigLabel(configObject, key) {
   if (typeof item === 'string') return item;
   return '';
 }
+
+/**
+ * Get the target user ID for spell data operations
+ * @returns {string} The user ID to use for spell data
+ * @private
+ */
+export function _getTargetUserId() {
+  let targetUserId = game.user.id;
+  if (game.user.isActiveGM) {
+    const actorOwner = game.users.find((user) => user.character?.id === this.actor.id);
+    if (actorOwner) targetUserId = actorOwner.id;
+    else log(2, `No owner found for actor ${this.actor.name}, using GM`);
+  }
+  return targetUserId;
+}
